refactor(api): use parcel and proof API
