+++ src/api/mod.rs
@@ -14,9 +14,9 @@ pub async fn serve(port: u16, shared: ShadowShared) -> std::io::Result<()> {
             .data(shared.clone())
             .service(web::resource("/version").to(root::version))
             .service(web::resource("/ethereum/count").route(web::get().to(ethereum::count)))
-            .service(web::resource("/ethereum/proposal").to(ethereum::proposal))
+            .service(web::resource("/ethereum/proof").to(ethereum::proposal))
             .service(web::resource("/ethereum/receipt/{tx}/{last}").to(ethereum::receipt))
-            .service(web::resource("/ethereum/header/{block}").to(ethereum::header))
+            .service(web::resource("/ethereum/parcel/{block}").to(ethereum::parcel))
     })
     .disable_signals()
     .bind(format!("0.0.0.0:{}", port))?
